% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{prints}
\alias{prints}
\alias{print.YABS}
\title{S3Methods for Printing}
\usage{
\method{print}{YABS}(x, ...)

}
\arguments{
\item{x}{Object from \code{YABS} package.}

\item{...}{Additional arguments}
}
\value{
Prints \code{YABS} object
}
\description{
Prints for \code{YABS} objects
}
\author{
Vithor Rosa Franco <vithorfranco@gmail.com>
}