% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarys.R
\name{summarys}
\alias{summarys}
\alias{summary.YABS_MCMC}
\title{S3Methods for summarying}
\usage{
\method{summary}{YABS_MCMC}(object, ...)

}
\arguments{
\item{object}{An \code{YABS_MCMC} object.}

\item{...}{Arguments to be passed to or from other methods.}
}
\value{
Summarise \code{YABS_MCMC} object
}
\description{
Summaries for \code{YABS_MCMC} objects
}
\author{
Vithor Rosa Franco <vithorfranco@gmail.com>
}