% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{prints}
\alias{prints}
\alias{print.YABS_MCMC}
\title{S3Methods for Printing}
\usage{
\method{print}{YABS_MCMC}(x, ...)

}
\arguments{
\item{x}{Object from \code{YABS} package.}

\item{...}{Additional arguments}
}
\value{
Prints \code{YABS_MCMC} object
}
\description{
Prints for \code{YABS_MCMC} objects
}
\author{
Vithor Rosa Franco <vithorfranco@gmail.com>
}